Emit example PowerShell code for client-side
This is not perfect, as nslookup makes additional DNS queries that result
in additional traffic to the dnsteal server and cause an extraneous zero
byte file to be created.  It would be nicer to use the Resolve-DnsName
cmdlet, but there appears to be no way of disabling the validation that
rejects some of the base64 encoding characters.

This currently does not support compression (-z) or single file transfers.